refacto + add main stuff
separate functions in different files

add a todo to remember stuff

add personal tests

change todo

change todo

add travis CI

add a single test for testing Travis

change name to trigger test

add README + travis Badge

update todo

add some tests

separate parsing into its own module + some fixes

update gitignore
