+++ .travis.yml
@@ -1,0 +1,7 @@
+language: python
+
+python:
+    - "3.6"
+
+script:
+    - pytest
