+++ README.md
@@ -1,0 +1,3 @@
+# Expert System
+
+[![Build Status](https://travis-ci.com/pablo-a/expert-system.svg?branch=pablo)](https://travis-ci.com/pablo-a/expert-system)
