+++ pablo_test/empty_facts
@@ -1,0 +1,3 @@
+A => B
+=
+?B
