+++ pablo_test/empty_lines
@@ -1,0 +1,9 @@
+# crash cuz add empty lines in tab
+A => B
+
+
+
+=A
+
+
+?B
