+++ pablo_test/multiline_facts
@@ -1,0 +1,8 @@
+A => B
+C => D
+
+=A
+=C
+
+?B
+?D
