+++ pablo_test/parentesis_incorrect
@@ -1,0 +1,3 @@
+)(A => B
+=A
+?B
