+++ pablo_test/trailing_whitespace
@@ -1,0 +1,6 @@
+A => B      
+B => C				
+
+=A		
+
+?B
