+++ tests/test_inputs.py
@@ -1,0 +1,5 @@
+def func_add(a, b):
+    return a + b
+
+def test_func():
+    assert func_add(1,2) == 3
